[step4] feature 신청 승인 된 사람들의 명수를 반환하도록 수정

+++ src/main/java/nextstep/courses/domain/session/SessionType.java
@@ -24,7 +24,7 @@ public class SessionType {
         return new SessionType(false, maxAttendance);
     }
 
-    public boolean canRegisterNewUser(int currentRegisteredUserCount) {
+    public boolean canRegisterNewUser(long currentRegisteredUserCount) {
         return isFree || currentRegisteredUserCount < maxAttendance;
     }
